Towards a clean build

+++ interfaces/acados_cpp/ocp_qp.cpp
@@ -60,7 +60,7 @@ ocp_qp::ocp_qp(std::vector<uint> nx, std::vector<uint> nu, std::vector<uint> nbx
 
     qp = std::unique_ptr<ocp_qp_in>(create_ocp_qp_in(dim.get()));
 
-    for (int stage = 0; stage <= N; ++stage) {
+    for (uint stage = 0; stage <= N; ++stage) {
         auto lbx = vector<double>(qp->dim->nbx[stage], -INFINITY);
         set("lbx", stage, lbx);
         auto ubx = vector<double>(qp->dim->nbx[stage], +INFINITY);
@@ -164,10 +164,10 @@ void ocp_qp::squeeze_dimensions() {
     vector<int> nbx;
     vector<vector<uint>> idxbx;
     vector<vector<double>> lower_boundx, upper_boundx;
-    for (int stage = 0; stage <= N; ++stage) {
+    for (uint stage = 0; stage <= N; ++stage) {
         vector<uint> idxb_stage;
         vector<double> lower_bound_stage, upper_bound_stage;
-        for (int idx = 0; idx < dimensions()["nbx"].at(stage); ++idx) {
+        for (uint idx = 0; idx < dimensions()["nbx"].at(stage); ++idx) {
             double lb = all_lbx.at(stage).at(idx), ub = all_ubx.at(stage).at(idx);
             if (lb != -INFINITY || ub != +INFINITY) {
                 // we have a double-sided bound at this index
@@ -186,10 +186,10 @@ void ocp_qp::squeeze_dimensions() {
     vector<int> nbu;
     vector<vector<uint>> idxbu;
     vector<vector<double>> lower_boundu, upper_boundu;
-    for (int stage = 0; stage <= N; ++stage) {
+    for (uint stage = 0; stage <= N; ++stage) {
         vector<uint> idxb_stage;
         vector<double> lower_bound_stage, upper_bound_stage;
-        for (int idx = 0; idx < dimensions()["nbu"].at(stage); ++idx) {
+        for (uint idx = 0; idx < dimensions()["nbu"].at(stage); ++idx) {
             double lb = all_lbu.at(stage).at(idx), ub = all_ubu.at(stage).at(idx);
             if (lb != -INFINITY || ub != +INFINITY) {
                 // we have a double-sided bound at this index
@@ -204,7 +204,7 @@ void ocp_qp::squeeze_dimensions() {
         idxbu.push_back(idxb_stage);
     }
     d_change_bounds_dimensions_ocp_qp(nbu.data(), nbx.data(), qp.get());
-    for (int stage = 0; stage <= N; ++stage) {
+    for (uint stage = 0; stage <= N; ++stage) {
         set("lbx", stage, lower_boundx.at(stage));
         set("ubx", stage, upper_boundx.at(stage));
         bounds_indices("x", stage, idxbx.at(stage));
@@ -226,12 +226,12 @@ void ocp_qp::expand_dimensions() {
     auto idxbx = bounds_indices("x");
 
     vector<vector<double>> lower_boundx, upper_boundx;
-    for (int stage = 0; stage <= N; ++stage) {
+    for (uint stage = 0; stage <= N; ++stage) {
         vector<double> lower_bound_stage, upper_bound_stage;
-        int bound_index = 0;
-        for (int state_idx = 0; state_idx < qp->dim->nx[stage]; ++state_idx) {
+        uint bound_index = 0;
+        for (uint state_idx = 0; state_idx < (uint)qp->dim->nx[stage]; ++state_idx) {
             double lb, ub;
-            if (bound_index < qp->dim->nbx[stage] && state_idx == idxbx.at(stage).at(bound_index)) {
+            if (bound_index < (uint)qp->dim->nbx[stage] && state_idx == idxbx.at(stage).at(bound_index)) {
                 lb = all_lbx.at(stage).at(bound_index);
                 ub = all_ubx.at(stage).at(bound_index);
                 ++bound_index;
@@ -254,12 +254,12 @@ void ocp_qp::expand_dimensions() {
     auto idxbu = bounds_indices("x");
 
     vector<vector<double>> lower_boundu, upper_boundu;
-    for (int stage = 0; stage <= N; ++stage) {
+    for (uint stage = 0; stage <= N; ++stage) {
         vector<double> lower_bound_control, upper_bound_control;
-        int bound_index = 0;
-        for (int control_idx = 0; control_idx < qp->dim->nu[stage]; ++control_idx) {
+        uint bound_index = 0;
+        for (uint control_idx = 0; control_idx < (uint)qp->dim->nu[stage]; ++control_idx) {
             double lb, ub;
-            if (bound_index < qp->dim->nbu[stage] && control_idx == idxbu.at(stage).at(bound_index)) {
+            if (bound_index < (uint)qp->dim->nbu[stage] && control_idx == idxbu.at(stage).at(bound_index)) {
                 lb = all_lbu.at(stage).at(bound_index);
                 ub = all_ubu.at(stage).at(bound_index);
                 ++bound_index;
@@ -276,7 +276,7 @@ void ocp_qp::expand_dimensions() {
 
     d_change_bounds_dimensions_ocp_qp(qp->dim->nu, qp->dim->nx, qp.get());
     
-    for (int stage = 0; stage <= N; ++stage) {
+    for (uint stage = 0; stage <= N; ++stage) {
         set("lbx", stage, lower_boundx.at(stage));
         set("ubx", stage, upper_boundx.at(stage));
         vector<uint> idx_states(qp->dim->nx[stage]);
